moocwatch/requirements.txt
Django==1.10.1
django-cors-headers==1.1.0
django-rest-auth==0.8.1
djangorestframework==3.4.6
psycopg2==2.6.2
six==1.10.0
defusedxml==0.4.1
django-allauth==0.27.0
oauthlib==2.0.0
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
